UI and ogic improvements

+++ app/src/main/java/com/example/gimnasio/data/model/MejorCargaCardio.kt
@@ -3,5 +3,5 @@ package com.example.gimnasio.data.model
 data class MejorCargaCardio(
     val tiempo: Int,
     val intensidad: Int,
-    val carga: Int
+    val carga: Float
 )
